pyproject.toml
[project]
name = "iomoments-oracle"
version = "0.0.0"
description = "Python numerical oracle and test harness for iomoments. Not a shipped artifact."
requires-python = ">=3.11"
license = { text = "AGPL-3.0-or-later" }
authors = [{ name = "Ed Hodapp", email = "ed@hodapp.com" }]

# Hard floor: Python 3.11 (matches global CLAUDE.md). Dev-only; nothing
# published to PyPI. scipy + numpy are the reference-math dependencies for
# the Pébay oracle; pytest / mypy / pylint / flake8 are the gate tools.
dependencies = [
    "numpy>=1.26",
    "scipy>=1.13",
    # pydantic v2 for the ontology schema (D009). Matches the
    # python_agent / fireasmserver lineage; pydantic v1 would force a
    # schema-model rewrite on every port.
    "pydantic>=2.6",
    # YAML as the human-authoring surface for the ontology source
    # (D009 Phase 5). The DAG persistence format stays JSON; YAML is
    # only the input to the builder.
    "pyyaml>=6.0",
]

[project.scripts]
build-iomoments-ontology = "iomoments_ontology.build:main"
audit-ontology = "audit_ontology.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "mypy>=1.11",
    "pylint>=3.2",
    "flake8>=7.1",
    # scipy ≥ 1.17 still doesn't ship inline stubs; scipy-stubs is the
    # community-maintained counterpart used by numpy-adjacent projects.
    "scipy-stubs>=1.17",
    # PyYAML's inline stubs live in the types-PyYAML package.
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

# Package layout: the Python code lives under tooling/src/, imported by
# short name (e.g., `iomoments_ontology`). Tests under tooling/tests/
# reference these imports directly; tests under tests/ cover the
# numerical oracle and don't depend on the ontology package.
[tool.setuptools]
package-dir = { "" = "tooling/src" }

[tool.setuptools.packages.find]
where = ["tooling/src"]

# ---------------------------------------------------------------------------
# pytest: branch coverage is mandatory. 100% is the aspirational target;
# configure the gate floor conservatively and tighten as the oracle matures.
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests", "tooling/tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--cov=tests",
    "--cov=iomoments_ontology",
    "--cov-branch",
    "--cov-report=term-missing",
]

[tool.coverage.run]
branch = true
# Coverage is measured on `tests/` (Python oracle — D005) and the
# `iomoments_ontology` package (the formal-requirements infrastructure
# — D009). When the C userland grows a Python frontend, add it here.
source = ["tests", "iomoments_ontology"]

[tool.coverage.report]
show_missing = true
skip_covered = false
# fail_under tightens as the oracle lands. Left unset today so the
# deliberately failing hello-world doesn't double-fail on coverage.

# ---------------------------------------------------------------------------
# mypy --strict. Global rule; no per-file escapes without an inline
# comment naming the concrete reason.
# ---------------------------------------------------------------------------
[tool.mypy]
strict = true
python_version = "3.11"
files = ["tests", "tooling/src", "tooling/tests"]
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
disallow_any_unimported = true
# The ontology package is installed as a namespace under tooling/src/;
# tell mypy where to look so it resolves intra-package imports.
mypy_path = "tooling/src"

# numpy/scipy ship their own stubs in current versions; if an older
# environment misses them, install types-* stubs rather than setting
# ignore_missing_imports globally.

# ---------------------------------------------------------------------------
# flake8 has no native pyproject support (as of flake8 7.x); it reads
# setup.cfg or .flake8. The .flake8 file at repo root carries the config.
# Kept here in a comment for discoverability.
# ---------------------------------------------------------------------------
# [tool.flake8]
# max-complexity = 5
# max-line-length = 79
